refactor(DEV-10629): add `<MoniteContextProvider/>` and scoping for the `<QueryClientProvider/>`
