+++ packages/sdk-react/src/utils/test-utils.tsx
@@ -84,6 +84,7 @@ export const Provider = ({
           monite,
           code: userLocale,
           sentryHub,
+          queryClient: client,
         }}
       >
         <I18nLocaleProvider
